// adobe-edge/src/internal/web/MediaEdge.d.ts
export interface paths {
// ...
  '/error': {
    /** @description Signals that an error has occurred */
    post: {
      requestBody?: {
        content: {
          'application/json': {
            events?: {
              xdm: {
                mediaCollection: {
                  playhead: number;
                  errorDetails: {
                    name: string;
                    /** @enum {string} */
                    source: 'player' | 'external';
                  };
                  /** @description The sessionID generated on sessionStart */
                  sessionID: string;
                  qoeDataDetails?: {
                    /** Format: int32 */
                    bitrate?: number;
                    /** Format: int32 */
                    droppedFrames?: number;
                    /** Format: int32 */
                    framesPerSecond?: number;
                    /** Format: int32 */
                    timeToStart?: number;
                  };
                };
                /** @default media.error */
                eventType: string;
                /** Format: date-time */
                timestamp: string;
              };
            }[];
          };
        };
      };
      responses: {
        /** @description No content */
        204: {
          content: never;
        };
        /** @description Bad request */
        400: {
          content: {
            'application/json': {
              type?: string;
              /** @default 400 */
              status?: number;
              title?: string;
              /** @description Error that caused the 400 status */
              detail?: string;
              report?: {
                requestId?: string;
                details?: {
                  /** @description Field that contains the error */
                  name?: string;
                  /** @description Error for that specific field */
                  reason?: string;
                }[];
              };
            };
          };
        };
        /** @description Not found */
        404: {
          content: {
            'application/json': {
              type?: string;
              /** @default 404 */
              status?: number;
              title?: string;
              detail?: string;
              report?: {
                requestId?: string;
                /** @description Error that caused the 404 status */
                details?: string;
              };
            };
          };
        };
      };
    };
  };
  '/pauseStart': {
    /** @description Sent when the user presses Pause. Because there is no resume event type, it is inferred when you send a play event after a pauseStart. */
    post: {
      requestBody?: {
        content: {
          'application/json': {
            events?: {
              xdm: {
                mediaCollection: {
                  playhead: number;
                  /** @description The sessionID generated on sessionStart */
                  sessionID: string;
                  qoeDataDetails?: {
                    /** Format: int32 */
                    bitrate?: number;
                    /** Format: int32 */
                    droppedFrames?: number;
                    /** Format: int32 */
                    framesPerSecond?: number;
                    /** Format: int32 */
                    timeToStart?: number;
                  };
                };
                /** @default media.pauseStart */
                eventType: string;
                /** Format: date-time */
                timestamp: string;
              };
            }[];
          };
        };
      };
      responses: {
        /** @description No content */
        204: {
          content: never;
        };
        /** @description Bad request */
        400: {
          content: {
            'application/json': {
              type?: string;
              /** @default 400 */
              status?: number;
              title?: string;
              /** @description Error that caused the 400 status */
              detail?: string;
              report?: {
                requestId?: string;
                details?: {
                  /** @description Field that contains the error */
                  name?: string;
                  /** @description Error for that specific field */
                  reason?: string;
                }[];
              };
            };
          };
        };
        /** @description Not found */
        404: {
          content: {
            'application/json': {
              type?: string;
              /** @default 404 */
              status?: number;
              title?: string;
              detail?: string;
              report?: {
                requestId?: string;
                /** @description Error that caused the 404 status */
                details?: string;
              };
            };
          };
        };
      };
    };
  };
  '/ping': {
    /** @description Use the Ping request during main content playback in cases when content must be sent every 10 seconds, regardless of other API events that have been sent. The first ping event should fire 10 seconds after main content playback has begun. For ad content, it must be sent every 1 second during ad tracking. */
    post: {
      requestBody?: {
        content: {
          'application/json': {
            events?: {
              xdm: {
                mediaCollection: {
                  playhead: number;
                  /** @description The sessionID generated on sessionStart */
                  sessionID: string;
                  qoeDataDetails?: {
                    /** Format: int32 */
                    bitrate?: number;
                    /** Format: int32 */
                    droppedFrames?: number;
                    /** Format: int32 */
                    framesPerSecond?: number;
                    /** Format: int32 */
                    timeToStart?: number;
                  };
                };
                /** @default media.ping */
                eventType: string;
                /** Format: date-time */
                timestamp: string;
              };
            }[];
          };
        };
      };
      responses: {
        /** @description No content */
        204: {
          content: never;
        };
        /** @description Bad request */
        400: {
          content: {
            'application/json': {
              type?: string;
              /** @default 400 */
              status?: number;
              title?: string;
              /** @description Error that caused the 400 status */
              detail?: string;
              report?: {
                requestId?: string;
                details?: {
                  /** @description Field that contains the error */
                  name?: string;
                  /** @description Error for that specific field */
                  reason?: string;
                }[];
              };
            };
          };
        };
        /** @description Not found */
        404: {
          content: {
            'application/json': {
              type?: string;
              /** @default 404 */
              status?: number;
              title?: string;
              detail?: string;
              report?: {
                requestId?: string;
                /** @description Error that caused the 404 status */
                details?: string;
              };
            };
          };
        };
      };
    };
  };
  '/play': {
    /** @description Sent when the player changes state to "playing" from another state, such as when the on ('Playing') callback is triggered by the player. Other states from which the player moves to "playing" include "buffering", when the user resumes from "paused", when the player recovers from an error, and during autoplay. */
    post: {
      requestBody?: {
        content: {
          'application/json': {
            events?: {
              xdm: {
                mediaCollection: {
                  playhead: number;
                  /** @description The sessionID generated on sessionStart */
                  sessionID: string;
                  qoeDataDetails?: {
                    /** Format: int32 */
                    bitrate?: number;
                    /** Format: int32 */
                    droppedFrames?: number;
                    /** Format: int32 */
                    framesPerSecond?: number;
                    /** Format: int32 */
                    timeToStart?: number;
                  };
                };
                /** @default media.play */
                eventType: string;
                /** Format: date-time */
                timestamp: string;
              };
            }[];
          };
        };
      };
      responses: {
        /** @description No content */
        204: {
          content: never;
        };
        /** @description Bad request */
        400: {
          content: {
            'application/json': {
              type?: string;
              /** @default 400 */
              status?: number;
              title?: string;
              /** @description Error that caused the 400 status */
              detail?: string;
              report?: {
                requestId?: string;
                details?: {
                  /** @description Field that contains the error */
                  name?: string;
                  /** @description Error for that specific field */
                  reason?: string;
                }[];
              };
            };
          };
        };
        /** @description Not found */
        404: {
          content: {
            'application/json': {
              type?: string;
              /** @default 404 */
              status?: number;
              title?: string;
              detail?: string;
              report?: {
                requestId?: string;
                /** @description Error that caused the 404 status */
                details?: string;
              };
            };
          };
        };
      };
    };
  };
// ...
}

export type webhooks = Record<string, never>;

export type components = Record<string, never>;

export type $defs = Record<string, never>;

export type external = Record<string, never>;